src/backend/requirements.txt
httpx
PyYAML
urllib3
anyio
requests
fastapi
pydantic
pydantic-settings
pydantic_core
pyjwt